Déplace un state afin d'avoir le nombre exact d'indicateurs sélectionnés dans le panneau latéral
